src/django/requirements.txt
Django==2.1
django-filter==2.0.0
django-rest-framework-roles==0.5
django-versatileimagefield==1.9
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
docutils==0.14
Markdown==2.6.11
psycopg2==2.7.5
python-dateutil==2.6.1
uWSGI==2.0.17

django-cors-headers==2.4.0
django-rest-knox==3.1.5
